requirements.txt
asgiref==3.3.4
Django==3.2.3
# Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter 